Make indel_impact a float

+++ src/cfeintact/get_indel_impact.py
@@ -3,7 +3,7 @@ from typing import Iterable, TypeVar
 T = TypeVar("T")
 
 
-def get_indel_impact(reference: Iterable[T], query: Iterable[T]) -> float:
+def get_indel_impact(reference: Iterable[T], query: Iterable[T]) -> int:
     shift = 0
     impacted = 0
     counter = 0
